pyproject.toml
[tool.poetry]
name = "social-media-api"
version = "0.1.0"
description = ""
authors = ["Arabov <yousefaldabbas0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
attrs = "23.1.0"
black = "23.11.0"
click = "8.1.7"
Django = "4.2.7"
django-debug-toolbar = "4.2.0"
django-filter = "23.3"
django-rest-framework = "0.1.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.26.5"
flake8 = "6.1.0"
inflection = "0.5.1"
jsonschema = "4.19.2"
jsonschema-specifications = "2023.11.1"
mccabe = "0.7.0"
mypy-extensions = "1.0.0"
packaging = "23.2"
pathspec = "0.11.2"
pillow = "10.1.0"
platformdirs = "4.0.0"
pycodestyle = "2.11.1"
pyflakes = "3.1.0"
PyJWT = "2.8.0"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
PyYAML = "6.0.1"
referencing = "0.31.0"
rpds-py = "0.12.0"
sqlparse = "0.4.4"
tomli = "2.0.1"
typing-extensions = "4.8.0"
uritemplate = "4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
Faker = "^26.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
extend-exclude = """
    /(
        .git
      | .hg
      | .mypy_cache
      | .tox
      | venv
      | _build
      | buck-out
      | build
      | dist
    )/
    """

[tool.isort]
profile = "black"
